# src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API documentation for the EMC Load Balancer Service.",
        "title": "EMC Load Balancer API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/products": {
            "get": {
                "description": "Get all products",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Products"
                ],
                "summary": "List products",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/emc_lb_src_pkg_res.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/emc_lb_src_pkg_res.APIResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new product (requires product:create permission)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Products"
                ],
                "summary": "Create product",
                "parameters": [
                    {
                        "description": "Product details",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/emc_lb_src_pkg_entities.CreateProductRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/emc_lb_src_pkg_res.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/emc_lb_src_pkg_res.APIResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/emc_lb_src_pkg_res.APIResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/products/{id}": {
            "get": {
                "description": "Get a single product by its ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Products"
                ],
                "summary": "Get product by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Product ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/emc_lb_src_pkg_res.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/emc_lb_src_pkg_res.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/emc_lb_src_pkg_res.APIResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Soft-delete a product (requires product:delete permission)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Products"
                ],
                "summary": "Delete product",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Product ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/emc_lb_src_pkg_res.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/emc_lb_src_pkg_res.APIResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/emc_lb_src_pkg_res.APIResponse"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update an existing product (requires product:update permission)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Products"
                ],
                "summary": "Update product",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Product ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Fields to update",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/emc_lb_src_pkg_entities.UpdateProductRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/emc_lb_src_pkg_res.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/emc_lb_src_pkg_res.APIResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/emc_lb_src_pkg_res.APIResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/shops": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new shop for the current user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Shops"
                ],
                "summary": "Create a new shop",
                "parameters": [
                    {
                        "description": "Shop details",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/emc_lb_src_pkg_entities.CreateShopRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/emc_lb_src_pkg_res.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/emc_lb_src_pkg_res.APIResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/emc_lb_src_pkg_res.APIResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/shops/my-shop": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get the shop details for the current user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Shops"
                ],
                "summary": "Get my shop",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/emc_lb_src_pkg_res.APIResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/emc_lb_src_pkg_res.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/emc_lb_src_pkg_res.APIResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "emc_lb_src_pkg_entities.CreateProductRequest": {
            "type": "object",
            "required": [
                "name",
                "price"
            ],
            "properties": {
                "allow_backorder": {
                    "type": "boolean"
                },
                "brand_id": {
                    "type": "string"
                },
                "category_id": {
                    "type": "string"
                },
                "description": {
                    "type": "string",
                    "maxLength": 2000
                },
                "images": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "is_featured": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 1
                },
                "original_price": {
                    "type": "number",
                    "minimum": 0
                },
                "price": {
                    "type": "number",
                    "minimum": 0
                },
                "shop_id": {
                    "type": "string"
                },
                "short_desc": {
                    "type": "string",
                    "maxLength": 500
                },
                "sku": {
                    "type": "string",
                    "maxLength": 100
                },
                "slug": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 1
                },
                "status": {
                    "type": "string",
                    "enum": [
                        "active",
                        "draft",
                        "inactive"
                    ]
                },
                "stock": {
                    "type": "integer",
                    "minimum": 0
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "thumbnail": {
                    "type": "string",
                    "maxLength": 2048
                }
            }
        },
        "emc_lb_src_pkg_entities.CreateShopRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "description": {
                    "type": "string",
                    "maxLength": 2000
                },
                "logo": {
                    "type": "string",
                    "maxLength": 2048
                },
                "name": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 3
                }
            }
        },
...
        "emc_lb_src_pkg_entities.UpdateProductRequest": {
            "type": "object",
            "properties": {
                "allow_backorder": {
                    "type": "boolean"
                },
                "brand_id": {
                    "type": "string"
                },
                "category_id": {
                    "type": "string"
                },
                "description": {
                    "type": "string",
                    "maxLength": 2000
                },
                "images": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "is_featured": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 1
                },
                "original_price": {
                    "type": "number",
                    "minimum": 0
                },
                "price": {
                    "type": "number",
                    "minimum": 0
                },
                "shop_id": {
                    "type": "string"
                },
                "short_desc": {
                    "type": "string",
                    "maxLength": 500
                },
                "sku": {
                    "type": "string",
                    "maxLength": 100
                },
                "slug": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 1
                },
                "status": {
                    "type": "string",
                    "enum": [
                        "active",
                        "draft",
                        "inactive"
                    ]
                },
                "stock": {
                    "type": "integer",
                    "minimum": 0
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "thumbnail": {
                    "type": "string",
                    "maxLength": 2048
                }
            }
        },
...
        "emc_lb_src_pkg_res.APIResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "$ref": "#/definitions/emc_lb_src_pkg_res.ErrorCode"
                },
                "data": {},
                "errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/emc_lb_src_pkg_res.FieldError"
                    }
                },
                "message": {
                    "type": "string"
                },
                "pagination": {},
                "status": {
                    "type": "integer"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "emc_lb_src_pkg_res.ErrorCode": {
            "type": "string",
            "enum": [
                "COMMON_BAD_REQUEST",
                "COMMON_NOT_FOUND",
                "COMMON_CONFLICT",
                "COMMON_INTERNAL_SERVER_ERROR",
                "COMMON_UNAUTHORIZED",
                "COMMON_FORBIDDEN",
                "COMMON_TOO_MANY_REQUESTS",
                "COMMON_INVALID_IDENTITY_TYPE"
            ],
            "x-enum-varnames": [
                "ErrCodeBadRequest",
                "ErrCodeNotFound",
                "ErrCodeConflict",
                "ErrCodeInternal",
                "ErrCodeUnauthorized",
                "ErrCodeForbidden",
                "ErrCodeTooManyRequests",
                "ErrCodeInvalidIdentityType"
            ]
        },
        "emc_lb_src_pkg_res.FieldError": {
            "type": "object",
            "properties": {
                "field": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}